More compime path fixes.
1. src/comptime/eval.cpp — the local parse_integer_literal parsed into
   int64_t, so a literal like uint64::max (18446744073709551615, used in
   std.limits.max's accumulator) overflowed and silently failed to
   parse. That made try_fold_comptime_only_call bail out with no
   diagnostic, leaving max[int8]() as an unresolvable call to a
   comptime-only instance the compiler never lowers — hence "call to
   max$int8 could not be resolved". Fixed by parsing into the unsigned
   bit pattern and reinterpreting as int64_t, consistent with how
   eval_cast already treats value::integer as a raw 64-bit pattern.
   2. src/semantic/check.cpp's materialize_const_literal wrote negative
   comptime results (e.g. min[int8]() == -128) as a signed decimal
   string into an int_lit token. int_lit text is never signed elsewhere
   in the compiler (negation is always a separate unary_expr), so
   bytecode_compiler/llvm_codegen's unsigned-only literal parsers choked
   on the - — "could not parse integer literal -128". Fixed by writing
   the two's-complement bit pattern instead, matching what every other
   literal path expects.

Also found while verifying: std/limits.kira's min[T]/max[T] had no
branch for float32/float64 at all — they fell through the integer
accumulator and returned 0/-1 for floats. Added explicit float branches
(±3.40282347e38 / ±1.7976931348623157e308) since this was silently
wrong, not just uncompiled.

+++ src/semantic/check.cpp
@@ -6834,8 +6834,18 @@ private:
     lit->span = span;
     switch (value.kind) {
     case comptime::value_kind::integer:
+      // `int_lit` token text is always an unsigned digit string — negative
+      // values are ordinarily spelled as a `unary_expr(neg, ...)` wrapping
+      // one (see `std.limits.min`'s own doc comment on this). A synthesized
+      // literal has no such wrapper, so a negative `value.integer` is
+      // written out as its 64-bit two's-complement bit pattern instead of a
+      // signed decimal string — `encode_literal`/`compile_literal_value`
+      // parse it back as raw bits and truncate to the target width exactly
+      // as they would for any other literal, so the resulting value is
+      // identical either way.
       lit->lit_kind = token_kind::int_lit;
-      lit->value = std::to_string(value.integer);
+      lit->value =
+          std::to_string(static_cast<std::uint64_t>(value.integer));
       break;
     case comptime::value_kind::floating:
       lit->lit_kind = token_kind::float_lit;
